// src/app/(root)/ctgom/interface.ts
export const ctgomSchema = z.object({
  localizacao: z.string().min(1, "Localização é obrigatória"),
  municipio_id: z.coerce.number("Selecione um município valido").min(1, "Município é obrigatório"),
    ano: z.string().regex(/^\d{4}-\d{2}-\d{2}$/, "Formato inválido (YYYY-MM-DD)"),
    uploads: z.array(z.object({
    nome: z.string().min(0, "Título é obrigatório"),
    tipo: z.string().min(0, "Descrição é obrigatória"),
    
  })).optional(),
  }

)